+++ flask_pluginkit/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-version = "1.3.0"
+version = "2.0.0"
 
 author = "staugur"
 
@@ -11,8 +11,8 @@ from .flask_pluginkit import PluginManager
 from .installer import PluginInstaller
 from .web import blueprint
 from .fixflask import Flask
-from .utils import BaseStorage, LocalStorage, RedisStorage
+from .utils import BaseStorage, LocalStorage, RedisStorage, PY2, string_types
 
 __all__ = ["Flask", "PluginManager", "PluginInstaller", "blueprint",
            "PluginError", "TarError", "ZipError", "InstallError", "CSSLoadError",
-           "BaseStorage", "LocalStorage", "RedisStorage"]
+           "BaseStorage", "LocalStorage", "RedisStorage", "PY2", "string_types"]